Fix BFloat16 crash and enable CUDA on ARM64
- common.py: Add .float() before .numpy() to handle bf16 tensors
  (PL downgrades 16-mixed to bf16-mixed on CPU, which crashes numpy)
- setup_spark.sh: Use PyTorch cu126 index for CUDA-enabled ARM64 wheels
  (default PyPI only has CPU-only wheels for aarch64)

+++ src/blue_frogs/models/common.py
@@ -77,8 +77,8 @@ class BaseClassifier(pl.LightningModule):
         return loss
 
     def on_validation_epoch_end(self):
-        all_probs = torch.cat([o["probs"] for o in self._val_outputs]).numpy()
-        all_labels = torch.cat([o["labels"] for o in self._val_outputs]).numpy()
+        all_probs = torch.cat([o["probs"] for o in self._val_outputs]).float().numpy()
+        all_labels = torch.cat([o["labels"] for o in self._val_outputs]).float().numpy()
         avg_loss = torch.stack([o["loss"] for o in self._val_outputs]).mean()
 
         if len(set(all_labels)) >= 2:
